youtube-reader/requirements.txt
psycopg2-binary
python-dotenv
ytmusicapi